elastiq-take-home-assessment/movie-sentiment-app/backend/requirements.txt
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
sqlalchemy>=2.0.23
cohere>=4.32
python-dotenv>=1.0.0
pydantic>=2.4.2
